Removing AR logic
This wasn't actually helping with AR at all, it was merely serving as
an overshoot limiter which was only ever doing anything on sharp
line-art.

Removing "AR" helps with continuity issues (sudden intensity changes)
and it also makes the model a bit sharper sometimes (almost always
imperceptible on real content).

It may come back if I change my mind but this is unlikely as I'd rather
improve the models by tweaking other things.