Adding chroma variant
This is literally the dumbest way of doing this. I'm just running The
same model once per chroma channel separately and then merging the result
at the end.

The correct way of doing this would be increasing the sizes of the input and
output to handle 3 channels simultaneously, which would allow the model to learn
how to extract information from the luma plane to reconstruct chroma.

Currently it's completely blind to luma, but I think it's probably already better
than everything else even at this reduced size.