Removing DIV2K variant from the repo.
People who want to use this for whatever reason can keep using it, but
I'm not going to be maintaining it in any way whatsoever.

I'm still not even convinced there's a meaningful difference between it
and its manga109 counterpart, and to be fair this repo is getting too
bloated.