Removing LL variants
The difference was very small and these pretty much only existed to please
distortion metrics when testing with content downsampled in linear light.

Only fansub encoders actively go out of their way to downscale in linear light,
and you should probably not be using ArtCNN on artificially downscaled content anyway.

I know this is not a good excuse but nnedi3, ravu and fsrcnnx are also trained with images
downscaled in gamma light, and they never offered LL variants.

This is, again, not an argument against LL downscaling in any way, it's just that
doing the inverse operation when going up might not necessarily be ideal.

These may or may not come back depending on how I feel about this in ~2 weeks.