Modernising infra
This migrates the code from TF2 to Keras 3 (using JAX as the backend). It also includes a new gen script.

This is all WIP and more changes are probably coming. As it stands the compute shaders aren't exactly different
from their fragment counterparts in the main branch, but they might perform a little better depending on the
hardware. C4F32 is roughly ~25% faster on my 6600XT running on Vulkan.

To do:
- Revisit training data.
- Revise which variants are worth keeping.
- Reduce the number of texture fetches / shader passes.
- Explore other performance optimisations (fp16, cooperative matrix multiply, etc)
- Start exporting ONNX models to be used with ML backends such as TRT (possibly faster for Nvidia users)