Replacing C4F64 with C16F64
This model will only be offered in the ONNX format for now,
it's intended mostly for people who want to use ArtCNN outside of mpv,
for encoding or image restoration tasks.

It generally scores in the same ballpark as waifu2x, but it should be
faster which is its major selling point as an alternative.

GLSL version is probably too slow for most people, but I'll think about
adding it later.