Adding ArtCNN_R8F64
This is an experimental model with the same architecture as the bigger
R16F96, but ~4x smaller.

R16F96 was found to be too aggressive sometimes which can make it
temporally unstable in the presence of noise or compression artifacts.

The intent here is that this smaller model trades reconstruction quality
for temporal stability.

This is only slightly better than the even smaller (and faster) C16F64
so perhaps its existence isn't justifiable.