Adding ArtCNN_R16F96B
This is the same as the normal R16F96 model but without biases.

The rationale is making the model luminance-equivariant in an attempt to
combat overfitting and sensitivity to noise.

There's research supporting that this is good for denoising models,
but whether this also works for SISR is unclear.

References:
https://arxiv.org/pdf/1906.05478
https://arxiv.org/pdf/2008.13751