Making R8F64 official
Since the feedback has been overwhelmingly positive there's no need to keep this as an experiment.

This is literally just a smaller version of R16F96 and its use-cases are the same. For most video
content I don't think the quality difference between them matters, but it gives you a 3x to 4x speed boost.