Promoting _SAFE models
The degradation model used in the _SAFE models seems to do better on real anime content
on average. Baseline used downscaled anime luma with many rotations and reflections, while
safe simply used luma as is.

For R8F64, this is the difference between the 2 variants on every single image available at
https://github.com/Light-Anime-Datasets/FHD-Anime:

| Model      | MAE        | PSNR        | SSIM       |
|------------|------------|-------------|------------|
| R8F64      | 0.00350296 | 45.16972628 | 0.98246216 |
| R8F64_SAFE | 0.00346839 | 45.25007072 | 0.98253319 |

I also asked people to try this out on shows that were previously problematic and it
seems to be an improvement in general.

I'm keeping the old models with the _UNSAFE suffix for now.