Adding Flax Notebooks
This is just something I put together while reading their getting started
documentation. I don't have any real intention of switching to Flax anytime
soon, but I felt like playing with it.

This seems to work just fine, but exporting ONNX from JAX is apparently
a pain in the ass that requires using TF as a bridge, and if I'm going
to do that I can also just keep using Keras.

Maybe I'll need this again in the future so I'll just keep it in the repo.

Might also try to find the PyTorch port while I'm at it.