Adding old PyTorch Notebooks
This is an old notebook I used in the past when I tried migrated the code
to PyTorch. It's not up to date with the changes to the model since then,
and I'm only pushing this to the repo to serve as a backup in the case I
need it in the future (you never know).

No, it doesn't have any shader gen code.